requirements.txt
streamlit
fpdf2
pillow
reportlab